requirements.txt
discord_webhook==0.17.0
pycryptodomex==3.15.0
PyYAML==6.0
asn1==2.7.0
Pillow==10.0.0
tqdm==4.65.0
tk==0.1.0
numpy==1.25.1